+++ NEWS.md
@@ -1,3 +1,12 @@
+# taxonbridge 1.0.3
+
+* Updated v1.0.2 documentation
+* Added `fuzzy_search()` method
+* Added `download_gbif()` method
+* Added set operation parameter to `get_inconsistencies()`
+* Fixed bug in `get_inconsistencies()` when uninomials are set to true
+* Added `download_ncbi()` method
+
 # taxonbridge 1.0.2
 
 * Added `get_taxa()` method